refactor: MainLayout.jsx を components/hooks/utils に分割（1807行→495行）
サイドバー/ヘッダー/モーダル等をcomponents/へ、
ウェイポイント操作/ポリゴン操作/衝突検知/キーボードショートカット等の
副作用ロジックをhooks/へ分離し、MainLayout.jsxをオーケストレータ化
