+++ src/components/MainLayout/utils/constants.js
@@ -1,0 +1,8 @@
+// MainLayout で使う定数群
+
+// デフォルトの中心位置: 東京タワー
+export const DEFAULT_CENTER = { lat: 35.6585805, lng: 139.7454329 }
+
+// 日本全国俯瞰用 (Issue #52)
+export const JAPAN_OVERVIEW_CENTER = { lat: 36.5, lng: 138.0 }
+export const JAPAN_OVERVIEW_ZOOM = 5
